tests and refactor
