+++ apps/web/package.json
@@ -9,6 +9,7 @@
     "lint": "eslint .",
     "format": "prettier --write \"**/*.{ts,tsx}\"",
     "typecheck": "tsc -b --noEmit",
+    "test": "vitest run",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -43,6 +44,7 @@
     "prettier-plugin-tailwindcss": "^0.7.2",
     "typescript": "catalog:",
     "typescript-eslint": "^8.57.1",
-    "vite": "catalog:"
+    "vite": "catalog:",
+    "vitest": "catalog:"
   }
 }
