+++ apps/web/src/lib/links.ts
@@ -1,0 +1,4 @@
+export const CLI_COMMAND = "npx plugins add supabase-community/supabase-plugin"
+export const DOCS_URL = "https://supabase.com/docs/guides/ai-tools/plugins"
+export const CREATE_PROJECT_URL = "https://supabase.com/dashboard/new"
+export const WEBSITE_URL = "https://supabase.com"
